Refactor: LogResponse 파일명 수정 #25
